Changed how ModLoader works. Reasoning in commit message.
ModLoader had no benefit from being an Object, and should not be treated
as an Object so it is now a singleton. 

The default position on access should be to restrict it, so the majority
of fields are now private, static and final. 

JavaDocs were added to everything. You all seriously need to write more
documentation.

loadMods now catches all the exceptions. Previously the exception was
caught higher up, but there are no benefits to doing this. 

System outputs are now funneled to the Logger which apparently doesn't
exist. From now on, every System.out.println call should be changed to
OpenModLoader.INSTANCE.LOGGER

ModLoader is now final, this means it can not be extended by other
classes. Due to how specific the ModLoader class is, it should not be
extended. This is part of defining intent of the code.